Created cards for every pet on the ourpets page and moved petcard to components folder
